Modify something according to liujunchen advices except codes related to DeviceType::kROCm, this will be surpported soon for merging to Master. All tests test and pass again
