+++ oneflow/core/device/cuda_util.cpp
@@ -33,7 +33,7 @@ limitations under the License.
 
 #ifdef WITH_ROCM
 
-#include "oneflow/core/ep/hip/cuda_stream.h"
+#include "oneflow/core/ep/rocm/cuda_stream.h"
 #include <hip/hip_runtime.h>
 
 #endif  // WITH_ROCM
