+++ oneflow/core/device/cuda_util.h
@@ -176,7 +176,7 @@ cudaError_t CudaDriverGetPrimaryCtxActive(int dev, int* active);
 #include <rccl.h>
 #include <hip/hip_fp16.h>
 #include "oneflow/core/device/cuda_pseudo_half.h"
-#include "oneflow/core/ep/hip/cuda_stream.h"
+#include "oneflow/core/ep/rocm/cuda_stream.h"
 #include "oneflow/core/hipdnn/hipdnn.h"
 
 // #if CUDA_VERSION >= 10020
