+++ oneflow/core/ep/rocm/cuda_device.cpp
@@ -13,9 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/ep/hip/cuda_device.h"
-#include "oneflow/core/ep/hip/cuda_event.h"
-#include "oneflow/core/ep/hip/cuda_stream.h"
+#include "oneflow/core/ep/rocm/cuda_device.h"
+#include "oneflow/core/ep/rocm/cuda_event.h"
+#include "oneflow/core/ep/rocm/cuda_stream.h"
 
 #ifdef WITH_ROCM
