+++ oneflow/core/ep/rocm/cuda_device_manager.cpp
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/ep/hip/cuda_device_manager.h"
+#include "oneflow/core/ep/rocm/cuda_device_manager.h"
 #include "oneflow/core/device/cuda_util.h"
 
 #ifdef WITH_ROCM
