+++ oneflow/core/ep/rocm/cuda_stream.cpp
@@ -13,13 +13,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/ep/hip/cuda_stream.h"
+#include "oneflow/core/ep/rocm/cuda_stream.h"
 #include "oneflow/core/job/global_for.h"
 #include "oneflow/core/job/resource_desc.h"
 #include "oneflow/core/hardware/node_device_descriptor_manager.h"
 #include "oneflow/core/hardware/cuda_device_descriptor.h"
-#include "oneflow/core/ep/hip/cuda_event.h"
-#include "oneflow/core/ep/hip/cuda_device.h"
+#include "oneflow/core/ep/rocm/cuda_event.h"
+#include "oneflow/core/ep/rocm/cuda_device.h"
 
 #ifdef WITH_ROCM
 
@@ -40,11 +40,6 @@ void SetAffinityByDevice(int dev_id) {
   if (!cuda_device) { return; }
   node_device_desc->Topology()->SetCPUAffinityByPCIBusID(cuda_device->PCIBusID());
   node_device_desc->Topology()->SetMemoryAffinityByPCIBusID(cuda_device->PCIBusID());
-}
-
-bool IsCuda9OnTuringDevice(const hipDeviceProp_t& prop) {
-  // return CUDA_VERSION >= 9000 && CUDA_VERSION < 9020 && prop.major == 7 && prop.minor == 5;
-  return false;
 }
 
 }  // namespace
@@ -93,30 +88,12 @@ CudaStream::CudaStream(CudaDevice* device)
   // cublas_handle
   OF_CUBLAS_CHECK(hipblasCreate(&cublas_handle_));
   OF_CUBLAS_CHECK(hipblasSetStream(cublas_handle_, cuda_stream_));
-// #if CUDA_VERSION >= 10010
-//   // cublas_lt_handle
-//   OF_CUBLAS_CHECK(cublasLtCreate(&cublas_lt_handle_));
-// #endif
-// #if CUBLAS_VERSION >= 11000
-//   if (ParseBooleanFromEnv("ONEFLOW_EP_CUDA_ENABLE_TF32_EXECUTION", true)) {
-//     OF_CUBLAS_CHECK(cublasSetMathMode(cublas_handle_, CUBLAS_TF32_TENSOR_OP_MATH));
-//   }
-// #endif  // CUBLAS_VERSION >= 11000
+
   workspace_size_ = kDefaultWorkspaceSize;
   OF_CUDA_CHECK(hipMalloc(&workspace_, workspace_size_));
-// #if CUBLAS_VERSION >= 11200
-//   OF_CUBLAS_CHECK(cublasSetWorkspace(cublas_handle_, workspace_, workspace_size_));
-// #endif  // CUBLAS_VERSION >= 11200
-  // cudnn_handle
-  // if (IsCuda9OnTuringDevice(device_properties())) {
-  //   OF_CUDA_CHECK(hipDeviceSynchronize());
-  //   OF_CUDA_CHECK(hipGetLastError());
-  // }
+
   OF_CUDNN_CHECK(hipdnnCreate(&cudnn_handle_));
-  // if (IsCuda9OnTuringDevice(device_properties())) {
-  //   OF_CUDA_CHECK(hipDeviceSynchronize());
-  //   hipGetLastError();
-  // }
+ 
   OF_CUDNN_CHECK(hipdnnSetStream(cudnn_handle_, cuda_stream_));
 }
 
@@ -125,9 +102,7 @@ CudaStream::~CudaStream() {
   OF_CUDA_CHECK(hipStreamSynchronize(cuda_stream_));
   OF_CUDNN_CHECK(hipdnnDestroy(cudnn_handle_));
   OF_CUBLAS_CHECK(hipblasDestroy(cublas_handle_));
-// #if CUDA_VERSION >= 10010
-//   OF_CUBLAS_CHECK(cublasLtDestroy(cublas_lt_handle_));
-// #endif
+
   OF_CUDA_CHECK(hipStreamDestroy(cuda_stream_));
   OF_CUDA_CHECK(hipFree(workspace_));
 }
@@ -161,10 +136,6 @@ void CudaStream::RecordEvent(Event* event) {
 hipStream_t CudaStream::cuda_stream() const { return cuda_stream_; }
 
 hipblasHandle_t CudaStream::cublas_handle() const { return cublas_handle_; }
-
-// #if CUDA_VERSION >= 10010
-// cublasLtHandle_t CudaStream::cublas_lt_handle() const { return cublas_lt_handle_; }
-// #endif
 
 void* CudaStream::cublas_workspace() const { return workspace_; }
