+++ oneflow/core/ep/rocm/cuda_stream.h
@@ -13,11 +13,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_EP_HIP_HIP_STREAM_H_
-#define ONEFLOW_CORE_EP_HIP_HIP_STREAM_H_
+#ifndef ONEFLOW_CORE_EP_ROCM_CUDA_STREAM_H_
+#define ONEFLOW_CORE_EP_ROCM_CUDA_STREAM_H_
 
 #include "oneflow/core/ep/include/stream.h"
-#include "oneflow/core/ep/hip/cuda_device.h"
+#include "oneflow/core/ep/rocm/cuda_device.h"
 
 #ifdef WITH_ROCM
 
@@ -165,4 +165,4 @@ class CudaStream : public Stream {
 
 #endif  // WITH_ROCM
 
-#endif  // ONEFLOW_CORE_EP_CUDA_CUDA_STREAM_H_
+#endif  // ONEFLOW_CORE_EP_ROCM_CUDA_STREAM_H_
