+++ oneflow/core/ep/rocm/primitive/add.hip.cpp
@@ -16,9 +16,9 @@ limitations under the License.
 */
 #include "oneflow/core/common/preprocessor.h"
 #include "oneflow/core/ep/include/primitive/add.h"
-#include "oneflow/core/ep/hip/primitive/type_seq.h"
+#include "oneflow/core/ep/rocm/primitive/type_seq.h"
 #include "oneflow/core/hip/elementwise.hip.h"
-#include "oneflow/core/ep/hip/cuda_stream.h"
+#include "oneflow/core/ep/rocm/cuda_stream.h"
 #include "oneflow/core/device/cuda_pseudo_bfloat16.h"
 
 namespace oneflow {
