+++ oneflow/core/ep/rocm/primitive/cast.hip.cpp
@@ -15,9 +15,9 @@ limitations under the License.
 */
 #include "oneflow/core/common/preprocessor.h"
 #include "oneflow/core/ep/include/primitive/cast.h"
-#include "oneflow/core/ep/hip/primitive/type_seq.h"
+#include "oneflow/core/ep/rocm/primitive/type_seq.h"
 #include "oneflow/core/hip/elementwise.hip.h"
-#include "oneflow/core/ep/hip/cuda_stream.h"
+#include "oneflow/core/ep/rocm/cuda_stream.h"
 
 namespace oneflow {
