+++ oneflow/core/ep/rocm/primitive/permute.hip.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 */
 #include "oneflow/core/ep/include/primitive/permute.h"
 #include "oneflow/core/ep/common/primitive/permute_impl.h"
-#include "oneflow/core/ep/hip/cuda_stream.h"
+#include "oneflow/core/ep/rocm/cuda_stream.h"
 #include <hip/hip_runtime.h>
 
 namespace oneflow {
